Correct mistakes to compute timestamp for each point and some logic in the code

+++ lslidar_c16/lslidar_c16_decoder/include/lslidar_c16_decoder/lslidar_c16_decoder.h
@@ -216,8 +216,8 @@ private:
     bool is_first_sweep;
     double last_azimuth;
     double sweep_start_time;
-    uint32_t sweep_start_time_nsec;
-    double packet_start_time;
+    unsigned long sweep_start_time_nsec;
+    uint32_t packet_start_time;
     int layer_num;
     Firing firings[FIRINGS_PER_PACKET];
